requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.3
SQLAlchemy==2.0.35
Werkzeug==3.0.4
gunicorn==23.0.0
